demo

+++ main.py
@@ -8,7 +8,9 @@ app = Flask(__name__)
 @app.route('/', methods = ['GET', 'POST'])
 def upload_file():
     if request.method == 'POST':
-        f = request.files['file']
+        f = request.files.get('file')
+        if f is None:
+        	return "No Image Uploaded!!!"
         f.save(secure_filename(f.filename))
         img = imread(secure_filename(f.filename))
         os.remove(secure_filename(f.filename))
@@ -19,6 +21,9 @@ def upload_file():
 @app.after_request
 def add_header(response):
     response.headers['X-UA-Compatible'] = 'IE=Edge,chrome=1,firefox=1'
+    response.headers['Cache-Control'] = "no-cache, no-store, must-revalidate"
+    response.headers["Pragma"] = "no-cache"
+    response.headers["Expires"] = "0"
     response.headers['Cache-Control'] = 'public, max-age=0'
     return response
